backend/iot_bookstore/requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.7
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.0
setuptools==77.0.3
sqlparse==0.5.3
tzdata==2025.1
